feat: fetch user data from user view to TopNavbar
